Split jwtmiddleware into multiple files
